+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Create `ricaun.Revit.Sdk.Sample` to use sdk `ricaun.Revit.Sdk`.
 - Create `ricaun.Revit.Sdk.Sample.Tests` to use sdk `ricaun.Revit.Sdk` and `ricaun.Revit.Sdk.Sample` reference.
 - Update `SolutionTemplates` to have `slnx` file.
+- Update `Templates` to support `TargetFrameworks` option.
 ### Tests
 - Create `ricaun.Revit.Templates.Tests` to install and build solution/project from templates.